+++ cmd/gomlx_checkpoints/metrics.go
@@ -3,8 +3,10 @@ package main
 import (
 	"flag"
 	"fmt"
+	"maps"
 	"path"
 	"regexp"
+	"slices"
 	"strings"
 
 	"github.com/charmbracelet/lipgloss"
@@ -23,14 +25,20 @@ var (
 
 func metrics(checkpointPaths, names []string) {
 	numCheckpoints := len(names)
-	checkpointPath := checkpointPaths[0]
 	_ = numCheckpoints
 
 	// Load metrics points from the checkpointPath.
-	trainingMetricsPath := path.Join(checkpointPath, plots.TrainingPlotFileName)
-	points := must.M1(plots.LoadPoints(trainingMetricsPath))
-	if len(points) == 0 {
-		klog.Errorf("No metrics found in %q", trainingMetricsPath)
+	points := make([][]plots.Point, numCheckpoints)
+	foundSomething := false
+	for ii, checkpointPath := range checkpointPaths {
+		trainingMetricsPath := path.Join(checkpointPath, plots.TrainingPlotFileName)
+		points[ii] = must.M1(plots.LoadPoints(trainingMetricsPath))
+		if len(points) > 0 {
+			foundSomething = true
+		}
+	}
+	if !foundSomething {
+		klog.Errorf("No metrics found in file %q in paths %v", plots.TrainingPlotFileName, checkpointPaths)
 	}
 
 	var metricsNamesMatcher *regexp.Regexp
@@ -49,69 +57,56 @@ func metrics(checkpointPaths, names []string) {
 			metricsTypes.Insert(name)
 		}
 	}
-	metricsUsed := types.MakeSet[string]()
 	nameToShort := make(map[string]string)
 	shortToName := make(map[string]string)
-	for _, point := range points {
-		nameToShort[point.MetricName] = point.Short
-		shortToName[point.Short] = point.MetricName
-		if metricsNamesMatcher != nil || metricsTypes != nil {
-			foundName := metricsNamesMatcher.MatchString(point.MetricName) || metricsNamesMatcher.MatchString(point.Short)
-			foundType := metricsTypes != nil && metricsTypes.Has(point.MetricType)
-			if !foundName && !foundType {
-				continue
+	metricsUsed := types.MakeSet[NameMetric]()
+	for modelIdx, pointsPerModel := range points {
+		for _, point := range pointsPerModel {
+			nameToShort[point.MetricName] = point.Short
+			shortToName[point.Short] = point.MetricName
+			if metricsNamesMatcher != nil || metricsTypes != nil {
+				foundName := metricsNamesMatcher.MatchString(point.MetricName) || metricsNamesMatcher.MatchString(point.Short)
+				foundType := metricsTypes != nil && metricsTypes.Has(point.MetricType)
+				if !foundName && !foundType {
+					continue
+				}
 			}
+			metricsUsed.Insert(NameMetric{names[modelIdx], point.Short})
 		}
-		metricsUsed.Insert(point.Short)
 	}
 
-	// map metric name to position in row, starting from 1 (position 0 is for the global set)
-	metricsOrder := make(map[string]int)
-	nextPos := 1
-	if *flagMetricsNames != "" {
-		// Start with the order given by the user:
-		for _, name := range strings.Split(*flagMetricsNames, ",") {
-			if short, found := nameToShort[name]; found {
-				name = short
-			}
-			if metricsUsed.Has(name) {
-				metricsOrder[name] = nextPos
-				nextPos++
-			}
+	// Map the metrics to the column number, starting from 1 (column 0 is for the global step)
+	metricsInOrder := slices.SortedFunc(maps.Keys(metricsUsed), func(a, b NameMetric) int {
+		if c := strings.Compare(a.Metric, b.Metric); c != 0 {
+			return c
 		}
-	}
-	// Append all other metric names, in alphabetical order.
-	for _, short := range xslices.SortedKeys(metricsUsed) {
-		if _, found := metricsOrder[short]; found {
-			continue // Already listed.
-		}
-		metricsOrder[short] = nextPos
-		nextPos++
+		return strings.Compare(a.Name, b.Name)
+	})
+	metricsOrder := make(map[NameMetric]int, len(metricsInOrder))
+	for idx, nameMetric := range metricsInOrder {
+		metricsOrder[nameMetric] = idx + 1
 	}
 
 	if *flagMetricsLabels {
-		ReportMetricsLabels(shortToName, nameToShort, metricsOrder)
+		ReportMetricsLabels(shortToName)
 	}
 
 	if *flagMetrics {
-		ReportMetrics(checkpointPath, metricsUsed, metricsOrder, points)
+		ReportMetrics(names, metricsOrder, points)
 	}
 }
 
+// NameMetric hold the pair of information.
+type NameMetric struct{ Name, Metric string }
+
 // ReportMetricsLabels list all metrics short and long names.
-func ReportMetricsLabels(shortToName map[string]string, nameToShort map[string]string, metricsOrder map[string]int) {
+func ReportMetricsLabels(shortToName map[string]string) {
 	fmt.Println(titleStyle.Render("Metrics Labels"))
 	table := newPlainTable(true, lipgloss.Center, lipgloss.Left)
 	table.Headers("Short", "MetricName")
-	rows := make([][]string, len(metricsOrder))
-	for short, idx := range metricsOrder {
-		name, found := shortToName[short]
-		if !found {
-			// metric manually selected by name:
-			name = short
-			short = nameToShort[name]
-		}
-		rows[idx-1] = []string{short, name}
+	rows := make([][]string, 0, len(shortToName))
+	for _, short := range xslices.SortedKeys(shortToName) {
+		rows = append(rows, []string{short, shortToName[short]})
 	}
 	for _, row := range rows {
 		table.Row(row...)
@@ -120,42 +115,78 @@ func ReportMetricsLabels(shortToName map[string]string, nameToShort map[string]s
 }
 
 // ReportMetrics of the model.
-func ReportMetrics(checkpointPath string, metricsUsed types.Set[string], metricsOrder map[string]int, points []plots.Point) {
-	fmt.Println(titleStyle.Render(fmt.Sprintf("Metrics %q", checkpointPath)))
+func ReportMetrics(names []string, metricsOrder map[NameMetric]int, points [][]plots.Point) {
+	numCheckpoints := len(names)
+	fmt.Println(titleStyle.Render("Metrics Table"))
 	table := newPlainTable(true, lipgloss.Right)
-	header := make([]string, 1+len(metricsUsed))
+	header := make([]string, 1+len(metricsOrder))
 	header[0] = "Global Step"
-	for name, idx := range metricsOrder {
-		header[idx] = name
+	for nameMetric, idx := range metricsOrder {
+		if numCheckpoints == 1 {
+			header[idx] = nameMetric.Metric
+		} else {
+			header[idx] = fmt.Sprintf("%s: %s", nameMetric.Name, nameMetric.Metric)
+		}
 	}
 	table.Headers(header...)
 
-	currentStep := int64(-1)
-	var currentRow []string
-	for _, point := range points {
-		step := int64(point.Step)
-		if step != currentStep {
-			if currentStep != -1 {
-				table.Row(currentRow...)
+	// Checks whether the indices reached on the end of all points.
+	pointsIndices := make([]int, numCheckpoints)
+	isFinished := func() bool {
+		for ii, points := range points {
+			if pointsIndices[ii] < len(points) {
+				return false
 			}
-			currentStep = step
-			currentRow = make([]string, 1+len(metricsUsed))
-			currentRow[0] = humanize.Comma(step)
 		}
-		idx, found := metricsOrder[point.Short]
-		if found {
-			var value string
-			switch point.MetricType {
-			case "accuracy":
-				value = fmt.Sprintf("%.2f%%", 100.0*point.Value)
-			default:
-				value = fmt.Sprintf("%.3g", point.Value)
+		return true
+	}
+
+	// Find the next global step to enumerate.
+	nextGlobalStep := func() int64 {
+		globalStep := int64(-1)
+		for modelIdx, pointsPerModel := range points {
+			if pointsIndices[modelIdx] < len(pointsPerModel) {
+				point := pointsPerModel[pointsIndices[modelIdx]]
+				if globalStep == -1 || int64(point.Step) < globalStep {
+					globalStep = int64(point.Step)
+				}
 			}
-			currentRow[idx] = value
 		}
+		return globalStep
 	}
-	if currentStep != -1 {
-		table.Row(currentRow...)
-	}
+
+	// Loop one global step at a time:
+	for !isFinished() {
+		currentGlobalStep := nextGlobalStep()
+		row := make([]string, 1+len(metricsOrder))
+		row[0] = humanize.Comma(currentGlobalStep)
+		for modelIdx, pointsPerModel := range points {
+			// Consume all points for the currentGlobalStep.
+			nameMetric := NameMetric{Name: names[modelIdx]}
+			for pointsIndices[modelIdx] < len(pointsPerModel) {
+				point := pointsPerModel[pointsIndices[modelIdx]]
+				if int64(point.Step) != currentGlobalStep {
+					// Nothing else in this globalStep for this model.
+					break
+				}
+				pointsIndices[modelIdx]++ // We'll consume this point, move the head forward.
+				nameMetric.Metric = point.Short
+				colIdx, found := metricsOrder[nameMetric]
+				if !found {
+					// We are not printing this metric.
+					continue
+				}
+				var value string
+				switch point.MetricType {
+				case "accuracy":
+					value = fmt.Sprintf("%.2f%%", 100.0*point.Value)
+				default:
+					value = fmt.Sprintf("%.3g", point.Value)
+				}
+				row[colIdx] = value
+			} // end for loop on all model points on currentGlobalStep.
+		} // end for loop on all models.
+		table.Row(row...)
+	} // end scanning all points.
 	fmt.Println(table.Render())
 }